Change package organization
