+++ app/src/androidTest/java/ch/epfl/sweng/SDP/home/HomeActivityTest.java
@@ -17,8 +17,8 @@ import org.junit.runner.RunWith;
 import ch.epfl.sweng.SDP.R;
 import ch.epfl.sweng.SDP.auth.Account;
 import ch.epfl.sweng.SDP.game.LoadingScreenActivity;
-import ch.epfl.sweng.SDP.game.drawing.DrawingOffline;
-import ch.epfl.sweng.SDP.game.drawing.DrawingOfflineItems;
+import ch.epfl.sweng.SDP.game.drawing.withoutItems.DrawingOffline;
+import ch.epfl.sweng.SDP.game.drawing.withItems.DrawingOfflineItems;
 import ch.epfl.sweng.SDP.localDatabase.LocalDbHandlerForAccount;
 import ch.epfl.sweng.SDP.shop.ShopActivity;
